Refactor dynamic integration type script and helper module
This commit refactors the "Dynamic Integration Type - Agent Smith.ps1" script and the "AgentSmithAPI.psm1" helper module.

In the script, the [Password] attribute for the $RegistrationSecret parameter is changed to [String]. This change allows the RegistrationSecret to be passed as a plain string instead of a secure password object. This will be addressed later in a more secure way.

Additionally, the script execution parameters have been tweaked to pass the exact ExternalID of the agent. This was done because I believe oddities of the inegration due to either platform should be solved in the platform with the oddity and not piled onto the functionality of one end.

In the helper module, there are no functional changes. The commit only fixes a trailing whitespace issue in the file.

These changes improve the readability and maintainability of the codebase.